Make imexam work again on NumPy 1 after #188! The checksum op had started failing due to some odd casting conditions on NumPy 1 and a missing dtype specifier in the np.add.reduce() call, compared with numdisplay's long-divergent copy of the code.

+++ pyraf/irafdisplay.py
@@ -150,7 +150,7 @@ class ImageDisplay:
         a = numpy.array([tid, thingct, subunit, 0, x, y, z, t],
                         dtype=int).astype(numpy.uint16)
         # Compute the checksum
-        sum = numpy.add.reduce(a)
+        sum = numpy.add.reduce(a, dtype=numpy.uint16)
         sum = 0xffff - (sum & 0xffff)
         a[3] = sum
         self._write(a.tobytes())
